feat: move FileSystemTestingTrait to a new library product FileSystemTesting
